test(webui-v2): share source fixture helper
