+++ crates/ironclaw_webui_v2/static/js/test-utils/source-for-test.mjs
@@ -1,0 +1,19 @@
+import { readFileSync } from "node:fs";
+
+export function sourceForTest(baseUrl, path, exportNames) {
+  const source = readFileSync(new URL(path, baseUrl), "utf8");
+  const lines = [];
+  let skippingImport = false;
+  for (const line of source.split("\n")) {
+    if (!skippingImport && line.startsWith("import ")) {
+      skippingImport = !line.trimEnd().endsWith(";");
+      continue;
+    }
+    if (skippingImport) {
+      skippingImport = !line.trimEnd().endsWith(";");
+      continue;
+    }
+    lines.push(line.replace(/^export function /, "function "));
+  }
+  return `${lines.join("\n")}\nglobalThis.__testExports = { ${exportNames.join(", ")} };`;
+}
